pyproject.toml
[tool.poetry]
name = "prediction"
version = "0.1.0"
description = ""
authors = ["Ferdina Kusumah <ferdina.kusumah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-auditlog = "^3.0.0"
tensorflow = "^2.16.1"
numpy = "1.26.4"
drf-yasg = "^1.21.7"
google-auth = "^2.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
